backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pymongo==4.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic[email]==2.10.3
